requirements.txt
numpy
pandas
matplotlib
openai
tiktoken
python-dotenv
